requirements/base.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
httpx==0.27.0
slowapi==0.1.9
google-cloud-storage==2.16.0
qrcode[pil]==7.4.2
python-multipart==0.0.9
